修正加法功能，实现大数加法

+++ main.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main() {
-    int a, b;
+    long long int a, b;
     cin >> a >> b;
     cout << a + b << endl;
     return 0;
